changes and bug fixes
- creating issues while logging in no longer have captcha
- all Id is now big int "migrate:fresh" to apply
-fix bug empty assignee when status is assigneed

+++ nmstrackersystem/database/migrations/2020_07_03_080409_add_assignee_id_to_issues.php
@@ -15,7 +15,7 @@ class AddAssigneeIdToIssues extends Migration
     {
         Schema::table('issues', function (Blueprint $table) {
             //
-            $table->integer('assignee_id')->nullable();
+            $table->bigInteger('assignee_id')->nullable();
         });
     }
