+++ nmstrackersystem/database/migrations/2020_07_05_074738_add_assignee__idd_to_issues.php
@@ -14,7 +14,7 @@ class AddAssigneeIddToIssues extends Migration
     public function up()
     {
         Schema::table('issues', function (Blueprint $table) {
-            $table->integer('assignee_Idd')->nullable();
+            $table->bigInteger('assignee_Idd')->nullable();
             //
         });
     }
